+++ src/CommandLineOptions.php
@@ -14,6 +14,7 @@ final readonly class CommandLineOptions
         public string $action,
         public array $ports,
         public ?int $replicaPort,
+        public ?string $generatedScriptPath,
         public int $replicas,
         public string $redisBinary,
         public string $redisCliBinary,